# apps/cli/templates/monorepo/apps/api/static/openapi.json
{
  "openapi": "3.0.2",
  "paths": {
...
    "/api/v1/users": {
      "get": {
        "description": "Retrieve a paginated list of Users that can be filtered, sorted, and preloaded.",
        "summary": "Get Many Users",
        "tags": [
          "user"
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "nullable": true
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "nullable": true
            }
          },
          {
            "name": "preloads",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "orderBy",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "orderDirection",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "operationId": "user.getMany",
        "responses": {
          "200": {
            "description": "200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "number",
                      "enum": [
                        200
                      ]
                    },
                    "message": {
                      "default": "Fetched paginated data successfully!",
                      "type": "string"
                    },
                    "success": {
                      "default": true,
                      "type": "boolean"
                    },
                    "total": {
                      "default": 946,
                      "type": "number"
                    },
                    "page": {
                      "type": "number"
                    },
                    "limit": {
                      "default": 20,
                      "type": "number"
                    },
                    "totalPages": {
                      "default": 48,
                      "type": "number"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "email": {
                            "type": "string",
                            "format": "email"
                          },
                          "username": {
                            "type": "string",
                            "minLength": 3,
                            "maxLength": 50
                          },
                          "googleId": {
                            "type": "string"
                          },
                          "emailVerifiedAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "lastLoginAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "isAdmin": {
                            "default": false,
                            "type": "boolean"
                          },
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "deletedAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        },
                        "required": [
                          "email",
                          "username",
                          "id",
                          "createdAt",
                          "updatedAt"
                        ]
                      }
                    }
                  },
                  "required": [
                    "status",
                    "page",
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "401",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "number",
                      "enum": [
                        401
                      ]
                    },
                    "message": {
                      "default": "Sorry, you are not authorized to access this resource.",
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    }
                  },
                  "required": [
                    "status",
                    "success"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "403",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "number",
                      "enum": [
                        403
                      ]
                    },
                    "message": {
                      "default": "Sorry, you do not have permission to access this resource.",
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    }
                  },
                  "required": [
                    "status",
                    "success"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "404",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "number",
                      "enum": [
                        404
                      ]
                    },
                    "message": {
                      "default": "The requested resource was not found.",
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    }
                  },
                  "required": [
                    "status",
                    "success"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "500",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "number",
                      "enum": [
                        500
                      ]
                    },
                    "message": {
                      "default": "Sorry, something went wrong on our end. Please try again later.",
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    }
                  },
                  "required": [
                    "status",
                    "success"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ]
      },
      "post": {
        "description": "Create a new User with the provided data, with validation and will return the created entity.",
        "summary": "Store User",
        "tags": [
          "user"
        ],
        "parameters": [],
        "operationId": "user.store",
        "requestBody": {
          "description": "Body",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "username": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 50
                  },
                  "googleId": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 8,
                    "maxLength": 128
                  }
                },
                "required": [
                  "email",
                  "username",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "201",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "email": {
                          "type": "string",
                          "format": "email"
                        },
                        "username": {
                          "type": "string",
                          "minLength": 3,
                          "maxLength": 50
                        },
                        "googleId": {
                          "type": "string"
                        },
                        "emailVerifiedAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "lastLoginAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "isAdmin": {
                          "default": false,
                          "type": "boolean"
                        },
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "deletedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      },
                      "required": [
                        "email",
                        "username",
                        "id",
                        "createdAt",
                        "updatedAt"
                      ]
                    },
                    "status": {
                      "default": 200,
                      "type": "integer"
                    },
                    "message": {
                      "default": "Request processed successfully.",
                      "type": "string"
                    },
                    "success": {
                      "default": true,
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "401",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "number",
                      "enum": [
                        401
                      ]
                    },
                    "message": {
                      "default": "Sorry, you are not authorized to access this resource.",
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    }
                  },
                  "required": [
                    "status",
                    "success"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "403",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "number",
                      "enum": [
                        403
                      ]
                    },
                    "message": {
                      "default": "Sorry, you do not have permission to access this resource.",
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    }
                  },
                  "required": [
                    "status",
                    "success"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "404",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "number",
                      "enum": [
                        404
                      ]
                    },
                    "message": {
                      "default": "The requested resource was not found.",
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    }
                  },
                  "required": [
                    "status",
                    "success"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "500",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "number",
                      "enum": [
                        500
                      ]
                    },
                    "message": {
                      "default": "Sorry, something went wrong on our end. Please try again later.",
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    }
                  },
                  "required": [
                    "status",
                    "success"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ]
      }
    },
...
    "/api/v1/users/{id}/restore": {
      "patch": {
        "description": "Restore a previously soft-deleted User by its ID and then return the restored entity.",
        "summary": "Restore User",
        "tags": [
          "user"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "preloads",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "operationId": "user.restore",
        "requestBody": {
          "description": "Body",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "default": 200,
                    "type": "integer"
                  },
                  "message": {
                    "default": "Request processed successfully.",
                    "type": "string"
                  },
                  "success": {
                    "default": true,
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "email": {
                          "type": "string",
                          "format": "email"
                        },
                        "username": {
                          "type": "string",
                          "minLength": 3,
                          "maxLength": 50
                        },
                        "googleId": {
                          "type": "string"
                        },
                        "emailVerifiedAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "lastLoginAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "isAdmin": {
                          "default": false,
                          "type": "boolean"
                        },
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "deletedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      },
                      "required": [
                        "email",
                        "username",
                        "id",
                        "createdAt",
                        "updatedAt"
                      ]
                    },
                    "status": {
                      "default": 200,
                      "type": "integer"
                    },
                    "message": {
                      "default": "Request processed successfully.",
                      "type": "string"
                    },
                    "success": {
                      "default": true,
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "401",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "number",
                      "enum": [
                        401
                      ]
                    },
                    "message": {
                      "default": "Sorry, you are not authorized to access this resource.",
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    }
                  },
                  "required": [
                    "status",
                    "success"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "403",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "number",
                      "enum": [
                        403
                      ]
                    },
                    "message": {
                      "default": "Sorry, you do not have permission to access this resource.",
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    }
                  },
                  "required": [
                    "status",
                    "success"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "404",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "number",
                      "enum": [
                        404
                      ]
                    },
                    "message": {
                      "default": "The requested resource was not found.",
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    }
                  },
                  "required": [
                    "status",
                    "success"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "500",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "number",
                      "enum": [
                        500
                      ]
                    },
                    "message": {
                      "default": "Sorry, something went wrong on our end. Please try again later.",
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    }
                  },
                  "required": [
                    "status",
                    "success"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ]
      }
    }
  },
  "info": {
    "version": "1.0.0",
    "title": "go-kickstart REST API - Documentation",
    "description": "go-kickstart REST API - Documentation"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "name": "access_token",
        "in": "cookie"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}